[#495] refactor: 불필요한 코드 삭제, 메서드 이동
